+++ kernel/include/arch/i386/interrupts/pit/pit.h
@@ -15,11 +15,11 @@
 //#include <arch/i386/interrupts/pit/timer.h>
 
 namespace PIT {
-    constexpr uint8_t kTimerFreq = 1193180 / 10000;
+    constexpr uint32_t kTimerFreq = 1193180 / 10000;
     constexpr uint16_t kPITChannel0 = 0x40;
     constexpr uint16_t kPITChannel3 = 0x43;
     constexpr uint8_t kPITInitCommand = 0x36;
 
     void Initialize(); //function to init the pit
-    void SetPITCount(uint8_t freq); //function to set the pit frequency
+    void SetPITCount(uint32_t freq); //function to set the pit frequency
 }
